Added Calulation to add new rows to gsheets before error
This modification calculates the total number of rows that will be needed in the sheet after the new data is added. If this number exceeds the current maximum rows in the sheet, it calls sh.add_rows() to expand the sheet before attempting the sh.update() operation.

This should prevent the "exceeds grid limits" error.